requirements.txt
requests
PyQt5
openai
google-generativeai
llama-cpp-python
pillow